just one function "getHost", which do not use config
