+++ src/api/controller/api/export.js
@@ -1,6 +1,6 @@
 import Koa from 'koa';
 import route from 'koa-route';
-import { host, cleanHost } from 'config';
+import { getHost, getCleanHost } from '../../../common/uris';
 import config from '../../../../config.json';
 
 import exporters from '../../exporters';
@@ -17,8 +17,8 @@ export const getExporter = type => {
 
 export const getExporterConfig = () => ({
     ...config,
-    host,
-    cleanHost,
+    host: getHost(),
+    cleanHost: getCleanHost(),
 });
 
 export async function exportFileMiddleware(
