+++ src/api/statements/LodexContext.js
@@ -2,6 +2,7 @@ import { MongoClient } from 'mongodb';
 import config from 'config';
 import set from 'lodash.set';
 import field from '../models/field';
+import { getHost, getCleanHost } from '../../common/uris';
 import publishedCharacteristic from '../models/publishedCharacteristic';
 
 export const LodexContext = MongoClientImpl =>
@@ -20,13 +21,11 @@ export const LodexContext = MongoClientImpl =>
         );
         const characteristics = await handlePublishedCharacteristic.findAllVersions();
         const fields = await handleField.findAll();
-        const host = config.host;
-        const cleanHost = config.cleanHost;
         set(data, `${target || '$context'}`, {
             fields,
             characteristics,
-            host,
-            cleanHost,
+            host: getHost(),
+            cleanHost: getCleanHost(),
         });
         feed.send(data);
         await handleDb.close();
